requirements.txt
meshtastic>=2.3.0
pypubsub>=4.0.3
